fix(events): sanitize queued system markers
